+++ src/testPlatforms/kotlin/dev/codebandits/container/gradle/helpers/setup.kt
@@ -3,6 +3,7 @@ package dev.codebandits.container.gradle.helpers
 import dev.codebandits.container.gradle.GradleProjectTest
 import java.io.File
 import java.nio.file.Path
+import java.util.*
 import kotlin.io.path.createDirectory
 
 internal fun GradleProjectTest.setupPluginLibsDir() {
@@ -10,6 +11,18 @@ internal fun GradleProjectTest.setupPluginLibsDir() {
   val libsDirectory = projectDirectory.resolve("libs").createDirectory()
   hostProjectRoot.resolve("build/libs/container-gradle-plugin.jar")
     .copyRecursively(libsDirectory.resolve("container-gradle-plugin.jar").toFile())
+  val metadataFile = hostProjectRoot.resolve("build/pluginUnderTestMetadata/plugin-under-test-metadata.properties")
+  check(metadataFile.exists()) { "metadata file not found: ${metadataFile.path}" }
+  val metadata = Properties().apply { metadataFile.inputStream().use { load(it) } }
+  val classpathValue = metadata.getProperty("implementation-classpath")
+    ?: error("implementation-classpath not found in ${metadataFile.path}")
+  val jarPaths = classpathValue.split(File.pathSeparator)
+  jarPaths.forEach { jarPath ->
+    val jarFile = File(jarPath)
+    if (jarFile.exists() && jarFile.extension == "jar") {
+      jarFile.copyRecursively(libsDirectory.resolve(jarFile.name).toFile(), overwrite = true)
+    }
+  }
 }
 
 internal fun Path.configureBuildGradlePluginFromLibsDir() {
@@ -18,7 +31,7 @@ internal fun Path.configureBuildGradlePluginFromLibsDir() {
     import dev.codebandits.container.gradle.tasks.ContainerRunTask
     buildscript {
       dependencies {
-        classpath files('libs/container-gradle-plugin.jar')
+        classpath(fileTree("libs") { include("*.jar") })
       }
     }
     apply plugin: 'dev.codebandits.container'
@@ -32,7 +45,7 @@ internal fun Path.configureBuildGradleKtsPluginFromLibsDir() {
     import dev.codebandits.container.gradle.tasks.ContainerRunTask
     buildscript {
       dependencies {
-        classpath(files("libs/container-gradle-plugin.jar"))
+        classpath(fileTree("libs") { include("*.jar") })
       }
     }
     apply(plugin = "dev.codebandits.container")